LayoutTests/inspector/debugger/debugger-step-into-event-listener-expected.txt

Tests that stepping into dispatchEvent() method will lead to a pause in the first event listener.

Recieved foo event(1)! Recieved foo event(2)!
Debugger was enabled.
Set timer for test function.
Stepping into...
Stepping into...
SUCCESS: Did step into event listener(fooEventHandler1).
Stepping into...
Stepping into...
Stepping into...
SUCCESS: Did step into event listener(fooEventHandler2).
Debugger was disabled.
